Refactor keys modal
