+++ frontend/src/components/buttonsRow/buttonsRow.tsx
@@ -1,26 +1,9 @@
-"use client";
-
-import { useState } from "react";
 import styles from "./buttonsRow.module.css";
-import { AddButton, SubmitButton } from "@/components/button/button";
-import Modal from "@/components/modal/modal";
-import SearchDropdown from "../searchDropdown/searchDropdown";
-import { SelectType } from "@/types/selectOptions";
-import axios from "@/axios/client";
-import User from "@/types/user";
+import { AddButton } from "@/components/button/button";
 import { useRouter } from "next/navigation";
 
-export default function ButtonRow({ users }: { users: User[] }) {
-  const [newKeyModal, setNewKeyModal] = useState(false);
-  const [selectedEmailOption, setSelectedEmailOption] =
-    useState<SelectType>(null);
-  const [disabled, setDisabled] = useState(false);
+export default function ButtonRow() {
   const router = useRouter();
-
-  const emails = users.map((user) => ({
-    label: user.email,
-    value: user.id.toString(),
-  }));
 
   const buttons = [
     {
@@ -34,8 +17,7 @@ export default function ButtonRow({ users }: { users: User[] }) {
       id: 2,
       name: "New Key",
       onClick: () => {
-        setDisabled(false);
-        setNewKeyModal(true);
+        router.push("/admin/newKeyModal");
       },
     },
     {
@@ -47,55 +29,11 @@ export default function ButtonRow({ users }: { users: User[] }) {
     },
   ];
 
-  const handleModalClose = () => {
-    setNewKeyModal(false);
-    setSelectedEmailOption(null);
-  };
-
-  const handleModalSubmit = async () => {
-    setDisabled(true);
-    if (newKeyModal && selectedEmailOption) {
-      // Send email for user to register a key
-      await axios.post(`/users/register`, {
-        // @ts-ignore
-        email: selectedEmailOption.label,
-      });
-    }
-
-    setSelectedEmailOption(null);
-    setNewKeyModal(false);
-    setDisabled(false);
-  };
-
-  const newKeyModalContent = (
-    <div>
-      <SearchDropdown
-        options={emails}
-        placeholder="Add Email..."
-        subheader="Email"
-        selectDropdown="tableModal"
-        setSelectedOption={setSelectedEmailOption}
-      />
-      <SubmitButton
-        disabled={disabled}
-        text="Submit"
-        onClick={handleModalSubmit}
-      />
-    </div>
-  );
-
   return (
     <div className={styles.buttonRowContainer}>
       {buttons.map(({ id, name, onClick }) => (
         <AddButton key={id} text={name} onClick={() => onClick()} />
       ))}
-      {newKeyModal && (
-        <Modal
-          closeModal={handleModalClose}
-          title="New Key"
-          content={newKeyModalContent}
-        />
-      )}
     </div>
   );
 }
